Fixed build errors, all tests passing.
Fixed (another) nasty casting bug in the
code that initializes the ept pointer. I
was using the old way of casting, and overlooking
that I was storing a host virtual pointer
inside a host physical pointer. This caused
invept to fail (invept checks if the eptp is
valid).

I also fixed invept and invvpid to print an
error message (like vmload and vmclear), rather
than generate an invalid opcode exception. It's
easier to debug and understand the problem.

In the future, we could more carefully parse the
error returned.