Installs PMFS filesystem.
This was done by applying pmfs.patch from our version
of the PMFS git repository (currently here -
https://gitlab.flux.utah.edu/xcap/pmfs), and squashing
all of the commits from the patch into this one.

pmfs.patch was generated by essentially doing a diff
on the v3.10-compatible PMFS (the head of the master
branch in our repo) and linux v3.10 (tagged in our
repo as linux-v3.10). There were a
few minor patch conflicts I had to manually fix
in the patch. Those changes are reflected in the
pmfs.patch that is in our repo.