libcap-integration: Re-factor kliblcd page alloc and mapping.
Mostly complete except for the bits that need the rb tree I'm
planning to put in place for translating physical addresses
to cptr's.

This may seem like silly refactoring, but it's cleaning up
and unifying a bunch of crap (including the more recent
feature for passing strings back and forth).