host-resource-trees: Update higher level alloc, map, unmap for trees.
To mirror what will happen inside an isolated container, when
code does a higher-level page alloc, it expects the pages to
be mapped in its virtual and physical address spaces. It also
expects to be able to do address -> cptr translations. This means
we need to insert the page memory object into the resource tree
during page alloc (and remove it when the pages are freed).

Similarly, for higher-level map/unmap, we need to update the resource
trees - as this is what will happen inside an isolated LCD, and
non-isolated code expects to be able to do address -> cptr
translation.

Almost done with this boring but important feature ...