host-resource-trees: Volunteer host memory.
When non-isolated code wants to "volunteer" host memory to the
microkernel's capability system, it invokes lcd_volunteer_pages,
lcd_volunteer_dev_mem, or lcd_volunteer_vmalloc_mem, depending on
the type of memory.

Internally, these check to see if the memory has already been
volunteered (so we don't get duplicates -- this is checked via the
global memory interval tree). If not, it inserts the memory into
the caller's cspace. The caller can subsequently share the memory
with e.g. an isolated LCD via the capability mediated interfaces.

I had support for this before, but there was no checking for
duplicate inserts (and this is a real problem with the pmfs example
for string sharing): Non-isolated code has no way of knowing
(without implementing data structures on its own) whether it inserted
host memory already or not, or whether some other non-isolated
code has.

Furthermore, now we have full support for address -> cptr translation
in the non-isolated side. This is also needed for the pmfs example
with string sharing: before, the non-isolated code just always
inserted memory every time to share strings, even if this lead
to duplicate inserts.

I think this is one of the "friction points" of embedding a
capability system inside a kernel: translation from host objects
to capabilities and back. For some objects, you can just embed
the cptr in the object itself (our "container structs"). But for
some things -- like memory -- it's not so easy. (For device memory,
the host kernel doesn't use a struct page to represent it. So we're
faced with creating our own giant array of data structures to
represent each page of device memory, and embedding the cptr in
that. Or instead -- as I have done -- use a data structure like
a tree to do a reverse lookup.)