Fix PAT initialization, shrink ioremap region for now.
PAT appears to work now with write back memory. Added an
arch check to ensure PAT is valid before VM entry.

The ioremap region was too big (the amount of metadata - bitmaps,
etc. - was too much). I shrunk it for now to 16 MBs. Later, we
can make it slightly larger.