liblcd-v2: Add missing syscalls, adjust liblcd syscall interfacing.
I'm using registers now instead of cheating with the utcb. This
is good in the end because it was getting ugly using the utcb
for syscall arguments (e.g., during a sync ipc send, the higher
level code needed to know that one of the utcb registers would
be used for the actual syscall - that's pretty bad).