test-v2: Add rb tree code to liblcd, fix unresolved mutex symbols.
There is a limitation with my preprocessor hacks: If a header
invokes a function inside another static inline function, I
can't modify the call site (that is the key) using a macro
undef/re-define. Instead, I need to just define the function
in another source file.

Another limitation (yet to be encountered) is a problematic
macro function invoked in a header. I can't re-define the
macro function before the call site using the post hooks. Will
cross that bridge when I get there.