test-v2: Update resource tree to use straight rb trees.
The design of resource trees has changed; see comments in
top of common/resource_tree.c. I'm allowing nodes with the
same intervals to be inserted, but different cptr's. This ugliness
is necessary for non-isolated land. I'm no longer using
interval trees for this reason, and because we aren't inserting
overlapping, non-equal intervals (this is the motivation for
interval trees).