stack-trace: Add interface to remember where struct module is.
We unload the kernel module from the host after we've dup'd it
and mapped it in the LCD. This means the struct module the host
uses goes bye-bye, and address->symbol resolution for addresses
in the LCD's .ko will return "not found" or garbage if some other
.ko was loaded in its place.

Therefore, we need a way for non-isolated code to remember where
the struct module copy is. We store this in the lcd_arch so that
it can man-handle it when it does stack tracing.