stack-trace: Stack trace 90 percent there. Wow, that was ugly.
I fought this thing until late in the morning. Since we're
working with a copy of the module image (.ko), some of the fields
in the struct module have to be updated (symbol and string table
fields). Then, everything works.

Still getting weird trace symbol for return address at the
bottom of the stack.