stack-trace: All works now!
Now seeing a full stack - including __init symbols and the top
return address (I do a check in dump stack to see if the top of
the stack has a return address - happens when we fault just before
setting up the stack frame).

Introduces another LCD-specific hack to the module loader. If we're
loading inside an LCD, we add the __init symbols to the symbol
and string tables that the loader builds (otherwise, the loader
just builds a permanent home for core symbols). It's all going
into core symtab and strtab, which is misleading since for LCDs
those tables will have both; but as the code states, those are just
temporary vars for the tables; eventually, symtab and strtab are
pointed to them (after the module's init has ran).