pmfs: mount: Fix deadlock in unregister fs.
unregister fs does a synchronize_rcu, which basically waits
for all cpu's to enter a "quiescent" state. The PMFS LCD thread
was never yielding as it waited for the response, so we had a
deadlock. I just needed to move cond_resched into the right place
in microkernel/run.c.