lcd/liblcd: Add netdev helpers
struct net_device instance of the driver is heavily modified by the kernel. In
order to keep things in sync, we need some internal helpers so as to avoid
volunteering memory for various structures and marshalling lists.These files
pull in a minimal set of helper functions from the kernel to keep things
intact.

Signed-off-by: Vikram Narayanan <vikram186@gmail.com>