lcd/ixgbe: Support timers
Timers can't run inside LCD as there is no support. So, start a timer in KLCD
(but with a lower frequency compared to the original one) and trigger a handler
inside LCD when the timer expires.

In case of ixgbe driver, a service_event is triggered by the timer.

Also add support for asynchronous register dump via sysfs module params.

Signed-off-by: Vikram Narayanan <vikram186@gmail.com>