lcd/microkernel: Patch relative addresses after load
LCDs module gets loaded onto the host, then moved to LCDs memory and
unloaded from the host. We need vmfunc instruction to be on the same
page (same virt addr) on both KLCD and LCD. Create a copy of LCDs vmfunc
page and use the same load address as KLCD's vmfunc page. Also, find the
rela section of vmfuncwrapper and patch the addresses to point to the
newly copied vmfunc page.