cptr: Increase size of bootstrap pages
We create more number of cptrs as we map more number of pages inside the
isolated domain. As a result, we run out of cptrs when hyperthreading is
enabled (more cpus = more cptrs).  Now, have 8 levels in the cptr tree
and hopefully we won't run out of cptrs. Increase the number of
bootstrap pages to accomodate 8 levels of cptrs.