liblcd: Add block layer functions
For interacting with the admin queue, we need to call out to map the
memory (blk_rq_map_kern) and make a subsequent call to (blk_execute_rq)
for executing the request, which inturn calls queue_rq function pointer
to submit the request to the driver. To avoid this back and forth
interactions during admin queue operations, bring a set of block layer
functions inside liblcd.