Fix a bug in lazy stack
Lazy stack cleanup relied on ffsl api which returns 0 if no bit is set
and returns 1 to N if bits are set from position 0 to n-1. This was not
taken into account during lazy stack cleanup.
Also replace the spin_locks with irqsave/restore version as pick_stack
can be called in both normal and irq contexts