kliblcd: move away from using current
While isolating bus drivers ioremapping and other stuff are not
triggered in the context of lvd boot, but at a later time.  This caused
a bunch of failures while booting xhci-hcd driver.  Move this to a
global now. It's a bit ugly as it assumes that there can be only one
KLCD thread.